generate rows instead of direct divs.
I tried to make the drawing surface divs (pixels)
resize to their container using flexbox. The
problem was that they'd only resize into their
flex direction, but not on the cross-axis. This
caused the divs to not spread along the cross
direction (vertically) and they'd all bunch-up
horizontally, leaving empty space at the bottom
of the container. I'm solving this by creating
extra div containers called "rows" that the
pixels sit in. Now flexbox can be used to size
the pixels in the horizontal directions, and the
rows will size the pixels in the vertical
direction.